spwflw/requirements.txt
nicegui>=1.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
ccxt>=4.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
aiohttp>=3.8.0
requests>=2.31.0
uvicorn>=0.23.0
fastapi>=0.100.0
# Alembic пригодится для миграций БД в будущем
alembic>=1.11.0